Change datetime selector to time selector
